tests/wpt/web-platform-tests/signed-exchange/sxg-invalid-validity-url.tentative.html

<!DOCTYPE html>
<title>Loading SignedHTTPExchange which validity URL is different origin from
request URL must fail</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN + '/signed-exchange/resources/sxg/sxg-invalid-validity-url.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location, innerURLOrigin() + '/signed-exchange/resources/inner-url.html');
  assert_true(message.is_fallback);
}, 'Loading SignedHTTPExchange which validity URL is different origin from ' +
   'request URL must fail');

</script>
</body>
